Reorganised and added new class names on the button
Added .btn--focus and .padding-left--1 to the button; the former to add the missing styles and the latter to ensure the border surrounds the entire box, which is not what it was doing originally. Removed .padding-left--1 from the parent div because I have moved this style to the button for aforementioned reasons.